Relocate components
