+++ src/components/ConnectionInfo.tsx
@@ -1,6 +1,6 @@
 import { useLobbyStore } from "@store/store";
 
-const ConnectStatus = function ConnectStatus() {
+export default function ConnectionInfo() {
   const isConnected = useLobbyStore((state) => state.isConnected);
   if (!isConnected) {
     return (
@@ -18,5 +18,3 @@ const ConnectStatus = function ConnectStatus() {
     return null;
   }
 };
-
-export default ConnectStatus;
