+++ src/components/Locations.tsx
@@ -1,7 +1,7 @@
 import Card from "@components/Card";
 import { useCrossedStore, useLobbyStore } from "@store/store";
 
-const Locations = function Locations() {
+export default function Locations() {
   const locations = useLobbyStore((state) => state.locations);
   const currentLocation = useLobbyStore((state) => state.currentLocation);
   const crossedLocations = useCrossedStore((state) => state.crossedLocations);
@@ -46,5 +46,3 @@ const Locations = function Locations() {
     return null;
   }
 };
-
-export default Locations;
