+++ src/components/Rules.tsx
