+++ src/components/layout/Header/Header.scss
@@ -1,4 +1,4 @@
-@import "../../styles/catppuccin";
+@import "../../../styles/catppuccin";
 
 .navbar-dark {
   background-color: $text;
