+++ src/components/modules/Chat/Chat.scss
@@ -1,4 +1,4 @@
-@import "../../styles/catppuccin";
+@import "../../../styles/catppuccin";
 
 .chat-box {
   background-color: $base !important;
