+++ src/components/modules/GameSettings/NewGameForm.scss
