+++ src/main.tsx
@@ -7,8 +7,8 @@ import "@fontsource/atkinson-hyperlegible";
 import "@fontsource/atkinson-hyperlegible/700.css";
 
 import Plausible from "plausible-tracker";
-import Header from "@components/Header/Header";
-import Footer from "@components/Footer/Footer";
+import Header from "@components/layout/Header/Header";
+import Footer from "@components/layout/Footer/Footer";
 const plausible = Plausible({
   domain: "spyfall.verybadfrags.com",
   apiHost: "/ps",
